Hubble/requirements.txt
pymongo~=4.6.0
pip==23.2.1
boto3==1.34.15
botocore
chardet>=3.0.2,<5;
bson
openai
cohere
numpy==1.24.3
pandas==2.0.3
psycopg2-binary
requests==2.31.0
Flask==2.2.2
werkzeug==2.2.2
torch==1.10.0
pytz~=2023.3.post1
safetensors==0.3.3
sentence-transformers==2.2.2
tokenizers==0.14
transformers==4.35.2
pytest~=7.4.0
python-dateutil~=2.8.2
requests-aws4auth
sshtunnel
tiktoken==0.5.1
sagemaker
werkzeug==2.2.2
python-dotenv
gnews
git+https://hubble-ec2-token:glpat-vJynSRiZ-YGZnynABtp1@gitlab.com/insight1565609/newspaper3k-extended.git#egg=newspaper3k
fuzzywuzzy
anthropic
python-Levenshtein